Make PIN + biometric mandatory; fix MainActivity; raw-string PIN
- Remove leftover in-app-update methods from MainActivity (compile fix).
- Mandatory auth setup: new AuthSetupScreen/AuthSetupViewModel + AuthSetupRoute.
  After onboarding the user must create a PIN and (when available) enable biometric
  before the app unlocks. New InitialStatus.SETUP_AUTH gates any seed-valid state
  that has no PIN into this flow.
- Onboarding completion now routes to the mandatory setup instead of straight home.
- Fix the PIN leading-zero bug: store/verify the raw digit string everywhere
  (AuthUseCase, SecuritySettingsViewModel, PinCodeConfigScreen, SecurityManager)
  instead of Int round-tripping.
- Simplify MyApplication logging (AppLogger already gates on debug).

Build: :app:assembleDebug green.

+++ app/src/main/java/com/pasich/encly/presentation/screen/PinCodeConfigScreen.kt
@@ -78,7 +78,7 @@ fun PinCodeConfigScreen(
                     if (currentInput == firstPin) {
                         animationState = PinAnimationState.SuccessAnimation
                         delay(1200)
-                        securityViewModel.activationPinAuth(firstPin.toInt())
+                        securityViewModel.activationPinAuth(firstPin)
                         navController.popBackStack()
                     } else {
                         errorText = "PIN-коди не збігаються. Спробуйте ще раз."
